test(pytorch): move test_import_gpu cases to single_gpu suite
test_import_gpu[transformer_engine] fails on the CPU-only default-runner
that unit-test runs on because TE 2.17 dlopens libcuda.so.1 at import
time; libcuda is only mounted into containers via nvidia-container-cli
when a GPU is present.

The test is correctly gated with @pytest.mark.skipif(not IS_CUDA) — that
gate is about the IMAGE variant (CUDA vs CPU), not about the test
runner's GPU availability. Moving both parametrized cases (flash_attn,
transformer_engine) into single_gpu/ where the workflow runs on
x86-g6xl-runner with --gpus all gives them a real driver mount at
import time.

flash_attn passed on the CPU fleet only because it defers libcuda dlopen
until the first CUDA call — the import itself succeeded. Moving it too
keeps GPU-only import checks in a single suite and avoids relying on
that lazy-loading behavior.

The IS_CUDA skip is preserved as a belt-and-suspenders guard for any
future CPU-variant coverage that reuses the single_gpu suite.

Unit suite retains COMMON_PACKAGES (torch, torchvision, torchaudio,
deepspeed, boto3, requests, yaml, packaging) — none of which need a
GPU driver at import.
